% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aMatrixSparseRevLex}
\alias{aMatrixSparseRevLex}
\title{Generates the row indices used internally to generate the sparse matrix A.}
\usage{
aMatrixSparseRevLex(bS, M)
}
\arguments{
\item{bS}{A binary matrix specifying the set of observation patterns. Each row encodes a single pattern.}

\item{M}{A vector of positive integers giving the alphabet sizes of the discrete variables.}
}
\value{
A vector of row indices.
}
\description{
Generates the row indices used internally to generate the sparse matrix A.
}
