% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colVector}
\alias{colVector}
\title{Generates the column indices used internally to generate the sparse matrix A.}
\usage{
colVector(cardS, cardChi)
}
\arguments{
\item{cardS}{The number of missingness patterns.}

\item{cardChi}{The cardinality of the full joint space.}
}
\value{
A vector of column indices.
}
\description{
Generates the column indices used internally to generate the sparse matrix A.
}
