% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infoS}
\alias{infoS}
\title{Calculates the total cardinality of the sample spaces.}
\usage{
infoS(bS, M)
}
\arguments{
\item{bS}{A binary matrix specifying the set of observation patterns. Each row encodes a single pattern.}

\item{M}{A vector of positive integers giving the alphabet sizes of the discrete variables.}
}
\value{
The total cardinality.
}
\description{
Calculates the total cardinality of the sample spaces.
}
