% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infoS2}
\alias{infoS2}
\title{Calculates the individual cardinalities of the sample spaces.}
\usage{
infoS2(bS, M)
}
\arguments{
\item{bS}{A binary matrix specifying the set of observation patterns. Each row encodes a single pattern.}

\item{M}{A vector of positive integers giving the alphabet sizes of the discrete variables.}
}
\value{
A vector of individual cardinalities.
}
\description{
Calculates the individual cardinalities of the sample spaces.
}
