% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{margProj}
\alias{margProj}
\title{Internal function multiplying a mass function by the sparse matrix A.}
\usage{
margProj(p, bS, M)
}
\arguments{
\item{p}{A subprobability mass function on the full joint space.}

\item{bS}{A binary matrix specifying the set of observation patterns. Each row encodes a single pattern.}

\item{M}{A vector of positive integers giving the alphabet sizes of the discrete variables.}
}
\value{
A collection of marginal mass functions.
}
\description{
Internal function multiplying a mass function by the sparse matrix A.
}
