% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcartest_package.R
\docType{package}
\name{mcartest-package}
\alias{mcartest}
\alias{mcartest-package}
\title{mcartest: Mean-based MCAR tests & bands}
\description{
Tools for testing the MCAR assumption in functional data.
}
\author{
\strong{Maintainer}: Lukas Stank \email{lukas.stank@campus.lmu.de}

}
\keyword{internal}
