% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionTest.R
\name{ProjectionTest}
\alias{ProjectionTest}
\title{Carry out a test of MCAR in a contingency table, given incomplete observations.}
\usage{
ProjectionTest(pSh, nS, bS, M, B)
}
\arguments{
\item{pSh}{A sequence of empirical mass functions calculated using incomplete observations.}

\item{nS}{A sequence of integers giving the numbers of incomplete observations used to calculate \code{pSh}.}

\item{bS}{A binary matrix specifying the set of observation patterns. Each row encodes a single pattern.}

\item{M}{A vector of positive integers giving the alphabet sizes of the discrete variables.}

\item{B}{An integer giving the number of bootstrap samples to be used to calibrate the test.}
}
\value{
The p-value the Monte Carlo test described in \insertCite{BS2022;textual}{MCARtest}.

The value of the test statistic \eqn{R()}.
}
\description{
Carry out a test of MCAR in a contingency table, given incomplete observations.
}
\examples{
bS=matrix(c(1,1,0, 1,0,1, 0,1,1),byrow=TRUE,ncol=3) # Our canonical 3d example
M=c(2,2,2)
nS=c(200,200,200)

pS=c(0.125,0.375,0.375,0.125,0.250,0.250,0.250,0.250,0.100,0.400,0.400,0.100)
P12=pS[1:4]; P13=pS[5:8]; P23=pS[9:12]
X12=t(rmultinom(1,size=nS[1],prob=P12)/nS[1])
X13=t(rmultinom(1,size=nS[2],prob=P13)/nS[2])
X23=t(rmultinom(1,size=nS[3],prob=P23)/nS[3])
pSh=cbind(X12,X13,X23)

ProjectionTest(pSh,nS,bS,M,99)
}
\references{
\insertRef{BS2022}{MCARtest}
}
