% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundErrors.R
\name{RoundErrors}
\alias{RoundErrors}
\title{Round errors in halfspace representations}
\usage{
RoundErrors(X, digits = 15)
}
\arguments{
\item{X}{A halfspace representation to be rounded.}

\item{digits}{An integer giving the number of significant figures to be kept.}
}
\value{
A rounded halfspace representation.
}
\description{
Round errors in halfspace representations
}
\examples{
bS=matrix(c(1,1,1,0, 1,0,0,1, 0,1,0,1, 0,0,1,1),byrow=TRUE,ncol=4)
RoundErrors("9007199254740992/6004799503160661") #Occurs in ConsMinkSumHrep(bS,c(2,2,2,2))



}
